ByteBuffer bug (and ajax) Turns out that javascript strings are annoying. string.charCodeAt(i) returns a potentially 2 byte value We needed to and this with 0xFF to make sure we just get on byte.
IN ADDITION TO THIS, we need to make sure that ajax requests
are expecting binary data.  By default, it expects ASCII data
which is 0-127.  We needed 128-255 too, and this would cause problems
By setting:
   xmlhttp.overrideMimeType('text\/plain; charset=x-user-defined');
this lets us get binary data back

This bug was found when the player ID exceeded 127.  Its fixed now